requirements/common.txt
django-axes==5.13.1
django-extensions==3.1.1
django-filter==2.4.0
django==3.1.7
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
psycopg2==2.8.6
python-dotenv==0.15.0